member-portal-react-flask/backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Mail==0.10.0
requests==2.32.3
bcrypt==4.2.0
python-dotenv==1.0.1
